pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shieldai-proxy"
version = "0.1.0"
description = "ShieldAI Security Proxy — reverse proxy with security middleware"
requires-python = ">=3.12"
license = {text = "MIT"}

dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "httpx>=0.27.0",
    "redis[hiredis]>=5.2.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.6.0",
    "pyyaml>=6.0",
    "structlog>=24.4.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx",
    "ruff>=0.8.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.setuptools.packages.find]
include = ["proxy*"]

[tool.ruff]
target-version = "py312"
line-length = 100
